Get first e2e test working

+++ src/ok_scoring/db/game_mapper.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Table, Integer, ForeignKey, String
-from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.dialects.postgresql import UUID, BIGINT
 
 from ok_scoring.db.metadata import metadata
 
@@ -8,7 +8,7 @@ game = Table(
     metadata,
     Column('key', UUID, primary_key=True),
     Column('description', String),
-    Column('date', Integer),
+    Column('date', BIGINT),
     Column('duration', Integer),
     Column('winningPlayerKey', UUID, ForeignKey('player.key'))
 )
